+++ src/components/index.tsx
@@ -3,3 +3,4 @@ export { default as Header } from "./Header";
 export { default as Input } from "./Input";
 export { default as Modal } from "./Modal";
 export { default as TextArea } from "./TextArea";
+export { default as Select } from "./Select";
