+++ src/layouts/main/index.tsx
@@ -1,4 +1,4 @@
-import Header from "@/components/common/Header";
+import { Header } from "@/components";
 import { Outlet } from "react-router-dom";
 import { MainWrapper, Main, MainContainer } from "./main.styled";
